requirements.txt
streamlit==1.28.0
PyPDF2==3.0.1
openai>=1.6.1,<2.0.0   
python-dotenv==1.0.0
langchain==0.0.350
langchain-openai==0.0.2
tiktoken==0.5.2
